requirements.txt
alembic==1.0.10
Click==7.0
Flask==1.0.3
Flask-Migrate==2.5.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
mysql-connector==2.2.9
Pillow==6.0.0
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.15.4
